# gateway/docs/swagger.yaml
basePath: /api/v1
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API gateway for the BEPetCare system, providing access
    to user, order, product, payment, appointment, and pet record services.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: BEPetCare Gateway API
  version: "1.0"
paths:
...
  /payments:
    post:
      consumes:
      - application/json
      description: Creates a new payment record with details like order ID, appointment
        ID, amount, description, and payment method
      parameters:
      - description: Payment details
        in: body
        name: request
        required: true
        schema:
          properties:
            amount:
              type: number
            appointment_id:
              type: integer
            description:
              type: string
            method:
              type: string
            order_id:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Payment created successfully
          schema:
            properties:
              payment_id:
                type: integer
            type: object
        "400":
          description: Invalid request or invalid payment method
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Create a new payment
      tags:
      - Payments
  /payments/{payment_id}:
    get:
      description: Retrieves a payment record using its unique ID
      parameters:
      - description: Payment ID
        in: path
        name: payment_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Payment details
          schema:
            properties:
              amount:
                type: number
              appointment_id:
                type: integer
              description:
                type: string
              method:
                type: string
              order_id:
                type: integer
              payment_id:
                type: integer
              status:
                type: string
            type: object
        "400":
          description: Payment ID is required or invalid format
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Payment not found
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get payment by ID
      tags:
      - Payments
  /payments/cancel:
    post:
      consumes:
      - application/json
      description: Cancels a payment link for a specific payment ID with a cancellation
        reason
      parameters:
      - description: Cancellation details
        in: body
        name: request
        required: true
        schema:
          properties:
            cancellation_reason:
              type: string
            payment_id:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Payment link cancelled successfully
          schema:
            properties:
              status:
                type: string
            type: object
        "400":
          description: Invalid request or invalid payment_id format
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Cancel a payment link
      tags:
      - Payments
  /payments/update-amount:
    put:
      consumes:
      - application/json
      description: Updates the amount for a specific payment ID
      parameters:
      - description: Payment amount update details
        in: body
        name: request
        required: true
        schema:
          properties:
            amount:
              type: number
            payment_id:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Payment amount updated successfully
          schema:
            properties:
              status:
                type: string
            type: object
        "400":
          description: Invalid request or invalid payment_id format
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Update payment amount
      tags:
      - Payments
  /payments/update-method:
    put:
      consumes:
      - application/json
      description: Updates the payment method for a specific payment ID
      parameters:
      - description: Payment method update details
        in: body
        name: request
        required: true
        schema:
          properties:
            method:
              type: string
            payment_id:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Payment method updated successfully
          schema:
            properties:
              status:
                type: string
            type: object
        "400":
          description: Invalid request, invalid payment_id format, or invalid payment
            method
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Update payment method
      tags:
      - Payments
  /payments/update-status:
    put:
      consumes:
      - application/json
      description: Updates the status of a payment record for a specific payment ID
      parameters:
      - description: Payment status update details
        in: body
        name: request
        required: true
        schema:
          properties:
            payment_id:
              type: string
            status:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Payment status updated successfully
          schema:
            properties:
              status:
                type: string
            type: object
        "400":
          description: Invalid request, invalid payment_id format, or invalid payment
            status
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Update payment status
      tags:
      - Payments
  /payments/url:
    post:
      consumes:
      - application/json
      description: Creates a payment URL for a specific payment with details like
        payment ID, amount, and description
      parameters:
      - description: Payment URL details
        in: body
        name: request
        required: true
        schema:
          properties:
            amount:
              type: number
            description:
              type: string
            payment_id:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Payment URL created successfully
          schema:
            properties:
              checkout_url:
                type: string
              payment_link_id:
                type: string
            type: object
        "400":
          description: Invalid request or invalid payment_id format
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Create a payment URL
      tags:
      - Payments
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"